+++ src/constraints/penetration.rs
@@ -2,6 +2,7 @@ use super::{Constraint, PositionConstraint};
 use crate::{collision::Collision, components::*, Vector};
 
 use bevy::prelude::*;
+use bevy::math::{DVec2, DVec3, DMat3, DQuat};
 
 /// A constraint between two bodies that prevents overlap with a given compliance.
 ///
@@ -14,11 +15,11 @@ pub struct PenetrationConstraint {
     pub entity2: Entity,
     pub collision_data: Collision,
     /// Lagrange multiplier for the normal force
-    pub normal_lagrange: f32,
+    pub normal_lagrange: f64,
     /// Lagrange multiplier for the tangential force
-    pub tangential_lagrange: f32,
+    pub tangential_lagrange: f64,
     /// The constraint's compliance, the inverse of stiffness, has the unit meters / Newton
-    pub compliance: f32,
+    pub compliance: f64,
     /// Normal force acting along this constraint
     pub normal_force: Vector,
 }
@@ -42,7 +43,7 @@ impl PenetrationConstraint {
         &mut self,
         body1: &mut RigidBodyQueryItem,
         body2: &mut RigidBodyQueryItem,
-        sub_dt: f32,
+        sub_dt: f64,
     ) {
         // Compute collision positions on the two bodies.
         // Subtracting the local center of mass from the position seems to be important for some shapes, although it's not shown in the paper. Something may be wrong elsewhere?
@@ -139,7 +140,7 @@ impl PenetrationConstraint {
         self.update_normal_force(n, sub_dt);
     }
 
-    fn update_normal_force(&mut self, normal: Vector, sub_dt: f32) {
+    fn update_normal_force(&mut self, normal: Vector, sub_dt: f64) {
         // Equation 10
         self.normal_force = self.normal_lagrange * normal / sub_dt.powi(2);
     }
